+++ packages/fx-core/src/common/constants.ts
@@ -77,5 +77,3 @@ export const GraphReadUserScopes = ["https://graph.microsoft.com/User.ReadBasic.
 export const SPFxScopes = (tenant: string) => [`${tenant}/Sites.FullControl.All`];
 export const AzureScopes = ["https://management.core.windows.net/user_impersonation"];
 export const AppStudioScopes = [`${getAppStudioEndpoint()}/AppDefinitions.ReadWrite`];
-
-export const AadSet: Set<string> = new Set<string>();
